+++ schema/sdss5db/catalogdb/lvm_magnitude/lvm_magnitude_copy.sql
@@ -1,0 +1,1 @@
+\copy catalogdb.lvm_magnitude from /uufs/chpc.utah.edu/common/home/sdss50/sdsswork/target/catalogs/lvm_magnitude.csv csv header;